Extra method to export for single lookups
The problem with the current .whois() method is that it can return results requesting the user to narrow down their search, however if the user passes in 'domain godaddy.com' it will fail the if(tldjs.tldExists(domain) === false) { throw new Error('Invalid a `domain name`'); } checkpoint. 

Not sure if any users would ever want results back from the whois lookup with spam domains and more steps, but likely asking for it to return with single results should be better.